requirements.txt
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=1.0.0
ecdsa>=0.18.0
filelock>=3.12.0
lmdb>=1.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pycryptodome>=3.18.0
cryptography>=41.0.0
pytest>=7.4.0
orjson>=3.9.0
